+++ Scroller/Scroller.js
@@ -215,22 +215,6 @@ Scroller.propTypes = /** @lends sandstone/Scroller.Scroller.prototype */ {
 	id: PropTypes.string,
 
 	/**
-	 * The initially hidden height of the vertical scrollbar.
-	 *
-	 * If a [`Header`]{@link sandstone/Panels.Header} and a `Scroller` are used inside
-	 * a [`Panel`]{@link sandstone/Panels.Panel} with `featureContent` prop set to true,
-	 * the `Header` will automatically collapse and the `Scroller`'s vertical scrollbar will
-	 * enlarge.
-	 *
-	 * This value would be the vertical scrollbar height difference between when the header collapses
-	 * and when the header expands.
-	 *
-	 * @type {Number}
-	 * @public
-	 */
-	initialHiddenHeight: PropTypes.number,
-
-	/**
 	 * Prevents scroll by dragging or flicking on the scroller.
 	 *
 	 * @type {Boolean}